debug test

+++ GLP/Examples/gspls/gspls-train-main.cpp
@@ -3,7 +3,24 @@
 //  GLP
 //
 //  Created by Zheng Shao on 11/23/15.
+//  Copyright (c) 2012 Saigo Laboratoire. All rights reserved.
 //
+//  This is free software with ABSOLUTELY NO WARRANTY.
+//
+//  This program is free software; you can redistribute it and/or modify
+//  it under the terms of the GNU General Public License as published by
+//  the Free Software Foundation; either version 2 of the License, or
+//  (at your option) any later version.
+//
+//  This program is distributed in the hope that it will be useful,
+//  but WITHOUT ANY WARRANTY; without even the implied warranty of
+//  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+//  GNU General Public License for more details.
+//
+//  You should have received a copy of the GNU General Public License
+//  along with this program; if not, write to the Free Software
+//  Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA
+//  02111-1307, USA
 //
 
 #include <cstdio>
@@ -51,7 +68,7 @@ int main(int argc, char* argv[])
     // let's begin
     train->timeStart();
     
-    int i = 0;
+    size_t i = 0;
     while (i < param->n) {
         cerr << "iter: " << i+1 << endl;
